+++ src/main/java/roomescape/member/dao/MemberDao.java
@@ -1,9 +1,12 @@
-package roomescape.member;
+package roomescape.member.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.member.domain.Member;
+
+import java.util.Optional;
 
 @Repository
 public class MemberDao {
@@ -27,9 +30,10 @@ public class MemberDao {
         return new Member(keyHolder.getKey().longValue(), member.getName(), member.getEmail(), "USER");
     }
 
-    public Member findByEmailAndPassword(String email, String password) {
-        return jdbcTemplate.queryForObject(
-                "SELECT id, name, email, role FROM member WHERE email = ? AND password = ?",
+    public Optional<Member> findByEmailAndPassword(String email, String password) {
+        String sql = "SELECT id, name, email, role FROM member WHERE email = ? AND password = ?";
+        return jdbcTemplate.query(
+                sql,
                 (rs, rowNum) -> new Member(
                         rs.getLong("id"),
                         rs.getString("name"),
@@ -37,7 +41,7 @@ public class MemberDao {
                         rs.getString("role")
                 ),
                 email, password
-        );
+        ).stream().findFirst();
     }
 
     public Member findByName(String name) {
@@ -51,5 +55,17 @@ public class MemberDao {
                 ),
                 name
         );
+    }
+
+    public Optional<Member> findById(Long id) {
+        String sql = "SELECT id, name, email, role FROM member WHERE id = ?";
+        return jdbcTemplate.query(sql, (rs, rowNum) ->
+                new Member(
+                        rs.getLong("id"),
+                        rs.getString("name"),
+                        rs.getString("email"),
+                        rs.getString("role")
+                ), id
+        ).stream().findFirst();
     }
 }
